refactor(cli): refactor(cli): split pre-launch.ts by concern (#250)
pre-launch.ts was a 586-line module mixing project detection, launch
prompts, and the indexing question. Split it into three focused files
under pre-launch/ and make pre-launch.ts a thin barrel re-export.

- pre-launch/project-check.ts: detectProjectKind, scaffoldAgentsMd,
  runProjectCheck, MANIFESTS, ProjectKind
- pre-launch/launch-prompts.ts: LaunchModeChoices, LaunchAbortedError,
  runLaunchPrompts, buildBadges, persistLaunchChoices
- pre-launch/indexing-question.ts: maybeAskAboutIndexing,
  countProjectFiles, resolveIndexThreshold, INDEXABLE_EXTS, SKIP_DIRS,
  DEFAULT_INDEX_QUESTION_THRESHOLD

No public API change — boot.ts and pre-launch.test.ts still import from
'./pre-launch.js'. Closes #250.

Verified: `tsc --noEmit` for @wrongstack/cli clean; vitest run on
tests/pre-launch.test.ts (59 tests) green.
